+++ TDEngine2/include/utils/Utils.h
@@ -513,27 +513,6 @@ namespace TDEngine2
 
 
 	/*!
-		\brief The method converts a given value of E_GRAPHICS_CONTEXT_GAPI_TYPE enumeration type into a string
-
-		\param[in] graphicsContextType A value of E_GRAPHICS_CONTEXT_GAPI_TYPE type
-
-		\return The method converts a given value of E_GRAPHICS_CONTEXT_GAPI_TYPE enumeration type into a string
-	*/
-
-	TDE2_API std::string GraphicsContextTypeToString(E_GRAPHICS_CONTEXT_GAPI_TYPE graphicsContextType);
-
-	/*!
-		\brief The method converts a given string value into corresponding value of E_GRAPHICS_CONTEXT_GAPI_TYPE type
-
-		\param[in] value A string which contains one of type of a graphics context including \"d3d11\", \"gl3x\", etc
-		
-		\return The method returns a value of E_GRAPHICS_CONTEXT_GAPI_TYPE type based on a given string one
-	*/
-
-	TDE2_API E_GRAPHICS_CONTEXT_GAPI_TYPE StringToGraphicsContextType(const std::string& value);
-
-
-	/*!
 		\brief The following macro is used to mark some features as unimplemented. The invocations
 		of the macro will cause interruption of execution's process
 	*/
@@ -591,27 +570,6 @@ namespace TDEngine2
 			*/
 
 			TDE2_API static E_FORMAT_TYPE GetFormatFromString(const std::string& str);
-	};
-
-
-	/*!
-		class CDeferOperation
-
-		\brief The class defines a scoped guard that invokes some operation
-		when it goes out of scope
-	*/
-
-	class CDeferOperation
-	{
-		public:
-			typedef std::function<void()> TCallbackType;
-		public:
-			TDE2_API CDeferOperation() = delete;
-			TDE2_API CDeferOperation(const TCallbackType& callback);
-
-			TDE2_API ~CDeferOperation();
-		private:
-			TCallbackType mCallback;
 	};