+++ tests/utils/CContainersTests.cpp
@@ -3,18 +3,3 @@
 #include <vector>
 
 using namespace TDEngine2;
-
-
-TEST_CASE("CDeferOperation Tests")
-{
-	SECTION("TestDeferExecution_PassFunction_ExecuteFunctionWhenGoesOutOfScope")
-	{
-		bool isEvaluated = false;
-
-		{
-			CDeferOperation defer{ [&isEvaluated]() { isEvaluated = true; } };
-		}
-
-		REQUIRE(isEvaluated);
-	}
-}